+++ src/main/kotlin/com/booleworks/kjobs/api/persistence/redis/RedisConfig.kt
@@ -12,6 +12,14 @@ interface RedisConfig {
     val jobPattern: String get() = "job:*"
     val heartbeatPattern: String get() = "heartbeat:*"
 
+    /**
+     * The limit that is used for the Redis scan operations.
+     *
+     * For most use cases the default value is good. For high performance use cases it can help to
+     * adjust the limit.
+     */
+    val scanLimit: Long get() = 1000L
+
     fun jobKey(uuid: String): String = "job:$uuid"
     fun inputKey(uuid: String): String = "input:$uuid"
     fun resultKey(uuid: String): String = "result:$uuid"