+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improved `RedisLongPollManager` performance by using a single Redis connection for publishing a completed job instead of creating a new one each time.
 - Improved `RedisPersistence` performance by re-using a couple of Redis connections instead of creating a new one each time.
 
+### Added
+- Configuration parameter `RedisConfig.scanLimit` to configure the limit of the Redis scan operations. The default is 1000 and has the same value as the previously hardcoded limit.
 
 ## [1.0.0-RC23] - 2025-05-07